backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
alembic==1.18.4
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
email-validator==2.3.0
python-multipart==0.0.22